Create Champion Chooser
Friend was asking me to pick a letter between 1 and 26 while playing League of Legends. He wanted to pick a champion that starts with the letter corresponding to the number I chose. This is a fun little program that automates that process. It picks the random letter, displays all the champions that start with that letter, and to take it a step further, it chooses a random champion from that list of champions.